Refactor common pipeline defaults into shared helper
Extract region validation, endpoint construction, and SigV4 auth setup
into newPipelineDefaults() so future collect sub-fields (otlp,
span_metrics) can reuse the same logic.
